Added scripts to run various experiments. Fixed bugs in attacks. Added a discretization to the adult data so that the linear attack can run on it

+++ attacks/NN_classifier.py
@@ -368,4 +368,4 @@ def nn_classification_reconstruction(cfg, synth, targets, known_features, hidden
         original_predictions = label_encoder.inverse_transform(predictions)
         reconstructed_targets[hidden_feature] = original_predictions
 
-    return reconstructed_targets.astype(int)
+    return reconstructed_targets
